requirements.txt
pandas==1.2.4
streamlit==0.88.0
seaborn==0.11.1
geopandas==0.9.0
numpy==1.19.5
folium==0.12.1
matplotlib==3.3.4
altair==4.1.0
streamlit_folium==0.4.0
Pillow==8.4.0
